+++ hutool-core/src/main/java/com/xiaoleilu/hutool/util/StrUtil.java
@@ -1500,11 +1500,7 @@ public class StrUtil {
 	 * @return 如果两个字符串相同，或者都是<code>null</code>，则返回<code>true</code>
 	 */
 	public static boolean equals(CharSequence str1, CharSequence str2) {
-		if (str1 == null) {
-			return str2 == null;
-		}
-
-		return str1.equals(str2);
+		return equals(str1, str2, false);
 	}
 
 	/**
@@ -1524,11 +1520,28 @@ public class StrUtil {
 	 * @return 如果两个字符串相同，或者都是<code>null</code>，则返回<code>true</code>
 	 */
 	public static boolean equalsIgnoreCase(CharSequence str1, CharSequence str2) {
+		return equals(str1, str2, true);
+	}
+	
+	/**
+	 * 比较两个字符串是否相等。
+	 * 
+	 * @param str1 要比较的字符串1
+	 * @param str2 要比较的字符串2
+	 * @param ignoreCase 是否忽略大小写
+	 * @return 如果两个字符串相同，或者都是<code>null</code>，则返回<code>true</code>
+	 * @since 3.2.0
+	 */
+	public static boolean equals(CharSequence str1, CharSequence str2, boolean ignoreCase) {
 		if (str1 == null) {
 			return str2 == null;
 		}
-
-		return str1.toString().equalsIgnoreCase(str2.toString());
+		
+		if(ignoreCase) {
+			return str1.toString().equalsIgnoreCase(str2.toString());
+		}else {
+			return str1.equals(str2);
+		}
 	}
 
 	/**